Remove orm dir and move contents to utils
