+++ packages/data-sql/src/query-converter/modifiers/target.ts
@@ -1,5 +1,5 @@
 import type { AbstractQueryTarget, AbstractQueryTargetNestedOne } from '@directus/data';
-import { createUniqueAlias } from '../../orm/create-unique-alias.js';
+import { createUniqueAlias } from '../../utils/create-unique-alias.js';
 import type {
 	AbstractSqlQueryFnNode,
 	AbstractSqlQueryJoinNode,
