+++ packages/data-sql/src/utils/index.ts
@@ -1,3 +1,4 @@
+export * from './create-unique-alias.js';
 export * from './get-mapped-queries-stream.js';
 export * from './map-result.js';
 export * from './numeric-operator-conversion.js';
